requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.12.1
anyio==4.0.0
arrow==1.3.0
async-lru==2.0.4
async-timeout==4.0.3
asyncodbc==0.1.1
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.5
azure-cosmos==4.5.1
azure-data-tables==12.4.4
azure-eventgrid==4.15.0
azure-functions==1.17.0
azure-identity==1.15.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.8.0
azure-keyvault-secrets==4.7.0
azure-servicebus==7.11.3
azure-storage-blob==12.18.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
deepdiff==6.6.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
inflection==0.5.1
isodate==0.6.1
Mako==1.2.4
MarkupSafe==2.1.3
msal==1.24.1
msal-extensions==1.0.0
multidict==6.0.4
mygeotab==0.8.8
numpy==1.26.1
ordered-set==4.1.0
orjson==3.9.10
pandas==2.1.2
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
pymssql==2.2.10
pyodbc==5.0.1
python-dateutil==2.8.2
python-rapidjson==1.13
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlalchemy==2.0.22
sqlanydb==1.0.13
twilio==8.10.0
types-python-dateutil==2.8.19.14
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
yarl==1.9.2
